refactor(CardList): move list of Cards to distinct component
