+++ src/components/CardList/components/index.tsx
@@ -1,0 +1,1 @@
+export {default as Cards} from './Cards';
